+++ src/App.js
@@ -1,4 +1,5 @@
 import './App.scss';
+import './typography.scss';
 import './utils/Notifications.scss';
 
 import { useEffect } from 'react';