+++ src/components/Editor/NewComponentButton/NewComponentButton.js
@@ -7,7 +7,7 @@ import Modal from 'react-modal';
 import { addProjectComponent, openFile, setNameError } from '../EditorSlice';
 import Button from '../../Button/Button'
 import NameErrorMessage from '../ErrorMessage/NameErrorMessage';
-import { CloseIcon, NewFileIcon } from '../../../Icons';
+import { CloseIcon, PlusIcon } from '../../../Icons';
 import { validateFileName } from '../../../utils/componentNameValidation';
 import { useTranslation } from 'react-i18next';
 
@@ -37,7 +37,7 @@ const NewComponentButton = () => {
 
     return (
       <div>
-        <Button buttonText={t('filePane.newFileButton')} ButtonIcon={NewFileIcon} buttonOuter onClickHandler={showModal} className="btn--primary btn--small proj-new-component-button" />
+        <Button buttonText={t('filePane.newFileButton')} ButtonIcon={PlusIcon} buttonOuter onClickHandler={showModal} className="btn--primary btn--small proj-new-component-button" />
 
         <Modal
           isOpen={modalIsOpen}